improved refacor in inventory actions removing friend class
