+++ src/source/NewUIInventoryActionController.cpp
@@ -1,4 +1,4 @@
-// NewUIInventoryActionController.cpp: implementation of the CNewUIInventoryActionController class.
+// NewUIInventoryActionController.cpp
 //////////////////////////////////////////////////////////////////////
 
 #include "stdafx.h"
@@ -17,23 +17,26 @@
 #include "DSPlaySound.h"
 #include "ZzzInterface.h"
 #include "UIManager.h"
+#include "ChangeRingManager.h"
+#include "PortalMgr.h"
+#include "CSQuest.h"
 
 namespace SEASON3B
 {
 
 CNewUIInventoryActionController::CNewUIInventoryActionController()
-    : m_pOwner(nullptr)
+    : m_pContext(nullptr)
 {
 }
 
-void CNewUIInventoryActionController::SetOwner(CNewUIMyInventory* pOwner)
+void CNewUIInventoryActionController::SetContext(IInventoryActionContext* pContext)
 {
-    m_pOwner = pOwner;
+    m_pContext = pContext;
 }
 
 bool CNewUIInventoryActionController::HandleInventoryActions(CNewUIInventoryCtrl* targetControl) const
 {
-    if (m_pOwner == nullptr || targetControl == nullptr)
+    if (m_pContext == nullptr || targetControl == nullptr)
     {
         return false;
     }
@@ -43,12 +46,12 @@ bool CNewUIInventoryActionController::HandleInventoryActions(CNewUIInventoryCtrl
         return HandlePickedItemPlacement(targetControl);
     }
 
-    if (m_pOwner->GetRepairMode() == CNewUIMyInventory::REPAIR_MODE_OFF && IsPress(VK_RBUTTON))
+    if (m_pContext->GetRepairMode() == REPAIR_MODE_OFF && IsPress(VK_RBUTTON))
     {
         return HandleRightClick(targetControl);
     }
 
-    if (m_pOwner->GetRepairMode() == CNewUIMyInventory::REPAIR_MODE_ON && IsPress(VK_LBUTTON))
+    if (m_pContext->GetRepairMode() == REPAIR_MODE_ON && IsPress(VK_LBUTTON))
     {
         return HandleRepairClick(targetControl);
     }
@@ -79,12 +82,12 @@ bool CNewUIInventoryActionController::HandlePickedItemPlacement(CNewUIInventoryC
 
     if (bFromInventorySystem)
     {
-        if (TryApplyJewel(targetControl, pPickedItem, pPickItem, iSourceIndex, iTargetIndex))
+        if (ApplyJewels(targetControl, pPickedItem, pPickItem, iSourceIndex, iTargetIndex))
         {
             return true;
         }
 
-        if (iTargetIndex != -1 && TryStackItem(targetControl, pPickItem, iSourceIndex, iTargetIndex))
+        if (iTargetIndex != -1 && TryStackItems(targetControl, pPickItem, iSourceIndex, iTargetIndex))
         {
             return true;
         }
@@ -98,35 +101,17 @@ bool CNewUIInventoryActionController::HandlePickedItemPlacement(CNewUIInventoryC
     return false;
 }
 
-bool CNewUIInventoryActionController::TryApplyJewel(CNewUIInventoryCtrl* targetControl,
-    CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
+bool CNewUIInventoryActionController::TryApplyJewel(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
 {
-    const bool bIsJewelType =
-        pPickItem->Type == ITEM_JEWEL_OF_BLESS
-        || pPickItem->Type == ITEM_JEWEL_OF_SOUL
-        || pPickItem->Type == ITEM_JEWEL_OF_LIFE
-        || pPickItem->Type == ITEM_JEWEL_OF_HARMONY
-        || pPickItem->Type == ITEM_LOWER_REFINE_STONE
-        || pPickItem->Type == ITEM_HIGHER_REFINE_STONE
-        || pPickItem->Type == ITEM_POTION + 160
-        || pPickItem->Type == ITEM_POTION + 161;
-
-    if (!bIsJewelType)
-    {
-        return false;
-    }
-
-    return CNewUIMyInventory::ApplyJewels(targetControl, pPickedItem, pPickItem, iSourceIndex, iTargetIndex);
+    return ApplyJewels(targetControl, pPickedItem, pPickItem, iSourceIndex, iTargetIndex);
 }
 
-bool CNewUIInventoryActionController::TryStackItem(CNewUIInventoryCtrl* targetControl,
-    ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
+bool CNewUIInventoryActionController::TryStackItem(CNewUIInventoryCtrl* targetControl, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
 {
-    return CNewUIMyInventory::TryStackItems(targetControl, pPickItem, iSourceIndex, iTargetIndex);
+    return TryStackItems(targetControl, pPickItem, iSourceIndex, iTargetIndex);
 }
 
-bool CNewUIInventoryActionController::TryMoveItem(CNewUIInventoryCtrl* targetControl,
-    CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
+bool CNewUIInventoryActionController::TryMoveItem(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
 {
     if (iTargetIndex < 0 || !targetControl->CanMove(iTargetIndex, pPickItem))
     {
@@ -148,12 +133,12 @@ bool CNewUIInventoryActionController::TryMoveItem(CNewUIInventoryCtrl* targetCon
 
 bool CNewUIInventoryActionController::HandleRepairClick(CNewUIInventoryCtrl* targetControl) const
 {
-    return CNewUIMyInventory::RepairItemAtMousePoint(targetControl);
+    return RepairItemAtMousePoint(targetControl);
 }
 
 bool CNewUIInventoryActionController::HandleRightClick(CNewUIInventoryCtrl* targetControl) const
 {
-    m_pOwner->ResetMouseRButton();
+    m_pContext->ResetMouseRButton();
 
     if (g_pNewUISystem->IsVisible(INTERFACE_STORAGE))
     {
@@ -197,8 +182,7 @@ bool CNewUIInventoryActionController::HandleStorageAutoMove(CNewUIInventoryCtrl*
 
 bool CNewUIInventoryActionController::HandleSellToNPC(CNewUIInventoryCtrl* targetControl) const
 {
-    CNewUIPickedItem* pExistingPickedItem = CNewUIInventoryCtrl::GetPickedItem();
-    if (pExistingPickedItem)
+    if (CNewUIInventoryCtrl::GetPickedItem())
     {
         return false;
     }
@@ -272,7 +256,7 @@ bool CNewUIInventoryActionController::HandleInventoryRightClickActions(CNewUIInv
 {
     if (g_pNewUISystem->IsVisible(INTERFACE_INVENTORY_EXT))
     {
-        return CNewUIMyInventory::TryTransferBetweenInventorySections(targetControl);
+        return TryTransferBetweenInventorySections(targetControl);
     }
 
     ITEM* pItem = targetControl->FindItemAtPt(MouseX, MouseY);
@@ -283,7 +267,7 @@ bool CNewUIInventoryActionController::HandleInventoryRightClickActions(CNewUIInv
 
     const int iIndex = targetControl->GetIndexByItem(pItem);
 
-    if (iIndex >= 0 && CNewUIMyInventory::TryConsumeItem(targetControl, pItem, iIndex))
+    if (iIndex >= 0 && TryConsumeItem(targetControl, pItem, iIndex))
     {
         return true;
     }
@@ -368,7 +352,7 @@ bool CNewUIInventoryActionController::TryEquipItem(CNewUIInventoryCtrl* targetCo
         return false;
     }
 
-    if (!m_pOwner->IsEquipable(nDstIndex, pItem))
+    if (!m_pContext->IsEquipable(nDstIndex, pItem))
     {
         return false;
     }
@@ -387,7 +371,7 @@ bool CNewUIInventoryActionController::TryEquipItem(CNewUIInventoryCtrl* targetCo
         }
     }
 
-    if (!m_pOwner->IsEquipable(nDstIndex, pItem))
+    if (!m_pContext->IsEquipable(nDstIndex, pItem))
     {
         return true;
     }
@@ -449,6 +433,603 @@ bool CNewUIInventoryActionController::TryDropItem(CNewUIInventoryCtrl* targetCon
 
     SocketClient->ToGameServer()->SendDropItemRequest(tx, ty, sourceIndex);
     SendDropItem = sourceIndex;
+
+    return true;
+}
+
+bool CNewUIInventoryActionController::RepairItemAtMousePoint(CNewUIInventoryCtrl* targetControl) const
+{
+    ITEM* pItem = targetControl->FindItemAtPt(MouseX, MouseY);
+    if (pItem == nullptr)
+    {
+        return true;
+    }
+
+    if (IsRepairBan(pItem))
+    {
+        return true;
+    }
+
+    const int iIndex = targetControl->GetIndex(pItem->x, pItem->y);
+
+    if (g_pNewUISystem->IsVisible(INTERFACE_NPCSHOP) && g_pNPCShop->IsRepairShop())
+    {
+        SocketClient->ToGameServer()->SendRepairItemRequest(iIndex, 0);
+    }
+    else
+    {
+        SocketClient->ToGameServer()->SendRepairItemRequest(iIndex, 1);
+    }
+
+    return true;
+}
+
+bool CNewUIInventoryActionController::ApplyJewels(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
+{
+    const bool bIsJewelType =
+        pPickItem->Type == ITEM_JEWEL_OF_BLESS
+        || pPickItem->Type == ITEM_JEWEL_OF_SOUL
+        || pPickItem->Type == ITEM_JEWEL_OF_LIFE
+        || pPickItem->Type == ITEM_JEWEL_OF_HARMONY
+        || pPickItem->Type == ITEM_LOWER_REFINE_STONE
+        || pPickItem->Type == ITEM_HIGHER_REFINE_STONE
+        || pPickItem->Type == ITEM_POTION + 160
+        || pPickItem->Type == ITEM_POTION + 161;
+
+    if (!bIsJewelType)
+    {
+        return false;
+    }
+
+    ITEM* pItem = targetControl->FindItem(iTargetIndex);
+    if (!pItem)
+    {
+        return false;
+    }
+
+    const int iType       = pItem->Type;
+    const int iLevel      = pItem->Level;
+    const int iDurability = pItem->Durability;
+
+    bool bSuccess = true;
+
+    if (iType > ITEM_WINGS_OF_DARKNESS
+        && iType != ITEM_CAPE_OF_LORD
+        && !(iType >= ITEM_WING_OF_STORM && iType <= ITEM_WING_OF_DIMENSION)
+        && !(ITEM_WING + 130 <= iType && iType <= ITEM_WING + 134)
+        && !(iType >= ITEM_CAPE_OF_FIGHTER && iType <= ITEM_CAPE_OF_OVERRULE)
+        && (iType != ITEM_WING + 135))
+    {
+        bSuccess = false;
+    }
+
+    if (iType == ITEM_BOLT || iType == ITEM_ARROWS)
+    {
+        bSuccess = false;
+    }
+
+    if ((pPickItem->Type == ITEM_JEWEL_OF_BLESS && iLevel >= 6)
+        || (pPickItem->Type == ITEM_JEWEL_OF_SOUL && iLevel >= 9))
+    {
+        bSuccess = false;
+    }
+
+    if (pPickItem->Type == ITEM_JEWEL_OF_BLESS
+        && iType == ITEM_HORN_OF_FENRIR
+        && iDurability != 255)
+    {
+        CFenrirRepairMsgBox* pMsgBox = nullptr;
+        CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CFenrirRepairMsgBoxLayout), &pMsgBox);
+        pMsgBox->SetSourceIndex(iSourceIndex);
+
+        const int iIndex = targetControl->GetIndex(pItem->x, pItem->y);
+        pMsgBox->SetTargetIndex(iIndex);
+
+        pPickedItem->HidePickedItem();
+        return true;
+    }
+
+    if (pPickItem->Type == ITEM_JEWEL_OF_HARMONY)
+    {
+        if (g_SocketItemMgr.IsSocketItem(pItem))
+        {
+            bSuccess = false;
+        }
+        else if (pItem->Jewel_Of_Harmony_Option != 0)
+        {
+            bSuccess = false;
+        }
+        else
+        {
+            const StrengthenItem strengthitem =
+                g_pUIJewelHarmonyinfo->GetItemType(static_cast<int>(pItem->Type));
+
+            if (strengthitem == SI_None)
+            {
+                bSuccess = false;
+            }
+        }
+    }
+
+    if (pPickItem->Type == ITEM_LOWER_REFINE_STONE
+        || pPickItem->Type == ITEM_HIGHER_REFINE_STONE)
+    {
+        if (g_SocketItemMgr.IsSocketItem(pItem))
+        {
+            bSuccess = false;
+        }
+        else if (pItem->Jewel_Of_Harmony_Option == 0)
+        {
+            bSuccess = false;
+        }
+    }
+
+    if (Check_LuckyItem(pItem->Type))
+    {
+        bSuccess = false;
+        if (pPickItem->Type == ITEM_POTION + 161)
+        {
+            if (pItem->Jewel_Of_Harmony_Option == 0) bSuccess = true;
+        }
+        else if (pPickItem->Type == ITEM_POTION + 160)
+        {
+            if (pItem->Durability > 0)               bSuccess = true;
+        }
+    }
+
+    if (bSuccess)
+    {
+        const int targetIndex = targetControl->GetIndexByItem(pItem);
+        SendRequestUse(iSourceIndex, targetIndex);
+        PlayBuffer(SOUND_GET_ITEM01);
+        return true;
+    }
+
+    return false;
+}
+
+bool CNewUIInventoryActionController::TryStackItems(CNewUIInventoryCtrl* targetControl, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const
+{
+    if (ITEM* pItem = targetControl->FindItem(iTargetIndex))
+    {
+        if (targetControl->AreItemsStackable(pPickItem, pItem))
+        {
+            SendRequestEquipmentItem(STORAGE_TYPE::INVENTORY, iSourceIndex,
+                pPickItem, STORAGE_TYPE::INVENTORY, iTargetIndex);
+            return true;
+        }
+    }
+
+    return false;
+}
+
+bool CNewUIInventoryActionController::TryConsumeItem(CNewUIInventoryCtrl* targetControl, ITEM* pItem, int iIndex) const
+{
+    if (pItem == nullptr)
+    {
+        return false;
+    }
+
+    if (pItem->Type == ITEM_TOWN_PORTAL_SCROLL)
+    {
+        SendRequestUse(iIndex, 0);
+        return true;
+    }
+
+    const auto isApple  = pItem->Type == ITEM_APPLE;
+    const auto isPotion =
+        (pItem->Type >= ITEM_APPLE && pItem->Type <= ITEM_ALE)
+        || (pItem->Type >= ITEM_SMALL_SHIELD_POTION && pItem->Type <= ITEM_LARGE_COMPLEX_POTION);
+
+    if (isApple || isPotion
+        || (pItem->Type == ITEM_POTION + 20 && pItem->Level == 0)
+        || (pItem->Type >= ITEM_JACK_OLANTERN_BLESSINGS && pItem->Type <= ITEM_JACK_OLANTERN_DRINK)
+        || (pItem->Type == ITEM_BOX_OF_LUCK && pItem->Level == 14)
+        || (pItem->Type >= ITEM_POTION + 70 && pItem->Type <= ITEM_POTION + 71)
+        || (pItem->Type >= ITEM_POTION + 72 && pItem->Type <= ITEM_POTION + 77)
+        || pItem->Type == ITEM_HELPER + 60
+        || pItem->Type == ITEM_POTION + 94
+        || (pItem->Type >= ITEM_CHERRY_BLOSSOM_WINE && pItem->Type <= ITEM_CHERRY_BLOSSOM_FLOWER_PETAL)
+        || (pItem->Type >= ITEM_POTION + 97 && pItem->Type <= ITEM_POTION + 98)
+        || pItem->Type == ITEM_HELPER + 81
+        || pItem->Type == ITEM_HELPER + 82
+        || pItem->Type == ITEM_POTION + 133)
+    {
+        SendRequestUse(iIndex, 0);
+        if (isApple)
+        {
+            PlayBuffer(SOUND_EAT_APPLE01);
+        }
+        else if (isPotion)
+        {
+            PlayBuffer(SOUND_DRINK01);
+        }
+
+        return true;
+    }
+
+    if (pItem->Type >= ITEM_POTION + 78 && pItem->Type <= ITEM_POTION + 82)
+    {
+        std::list<eBuffState> secretPotionbufflist;
+        secretPotionbufflist.push_back(eBuff_SecretPotion1);
+        secretPotionbufflist.push_back(eBuff_SecretPotion2);
+        secretPotionbufflist.push_back(eBuff_SecretPotion3);
+        secretPotionbufflist.push_back(eBuff_SecretPotion4);
+        secretPotionbufflist.push_back(eBuff_SecretPotion5);
+
+        if (g_isCharacterBufflist((&Hero->Object), secretPotionbufflist) == eBuffNone)
+        {
+            SendRequestUse(iIndex, 0);
+            return true;
+        }
+
+        CreateOkMessageBox(GlobalText[2530], RGBA(255, 30, 0, 255));
+        return false;
+    }
+
+    if ((pItem->Type >= ITEM_HELPER + 54 && pItem->Type <= ITEM_HELPER + 57)
+        || (pItem->Type == ITEM_HELPER + 58
+            && gCharacterManager.GetBaseClass(Hero->Class) == CLASS_DARK_LORD))
+    {
+        WORD point[5] = { 0, };
+        point[0] = CharacterAttribute->Strength  + CharacterAttribute->AddStrength;
+        point[1] = CharacterAttribute->Dexterity + CharacterAttribute->AddDexterity;
+        point[2] = CharacterAttribute->Vitality  + CharacterAttribute->AddVitality;
+        point[3] = CharacterAttribute->Energy    + CharacterAttribute->AddEnergy;
+        point[4] = CharacterAttribute->Charisma  + CharacterAttribute->AddCharisma;
+
+        wchar_t nStat[MAX_CLASS][5] =
+        {
+            18, 18, 15, 30,  0,
+            28, 20, 25, 10,  0,
+            22, 25, 20, 15,  0,
+            26, 26, 26, 26,  0,
+            26, 20, 20, 15, 25,
+            21, 21, 18, 23,  0,
+            32, 27, 25, 20,  0,
+        };
+
+        const int attributeType   = pItem->Type - (ITEM_HELPER + 54);
+        const int characterClass  = gCharacterManager.GetBaseClass(Hero->Class);
+        point[attributeType]     -= nStat[characterClass][attributeType];
+
+        if (point[attributeType] < (pItem->Durability * 10))
+        {
+            g_pMyInventory->SetStandbyItemKey(pItem->Key);
+            CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUsePartChargeFruitMsgBoxLayout));
+            return false;
+        }
+
+        SendRequestUse(iIndex, 0);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 58
+        && gCharacterManager.GetBaseClass(Hero->Class) != CLASS_DARK_LORD)
+    {
+        CreateOkMessageBox(GlobalText[1905]);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_ARMOR_OF_GUARDSMAN)
+    {
+        if (IsUnitedMarketPlace())
+        {
+            wchar_t szOutputText[512];
+            mu_swprintf(szOutputText, L"%ls %ls", GlobalText[3014], GlobalText[3015]);
+            CreateOkMessageBox(szOutputText);
+            return true;
+        }
+
+        if (Hero->SafeZone == false)
+        {
+            CreateOkMessageBox(GlobalText[2330]);
+            return false;
+        }
+
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(4, pItem->Level);
+        g_pMyInventory->SetStandbyItemKey(pItem->Key);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 46)
+    {
+        const BYTE byPossibleLevel = CaculateFreeTicketLevel(FREETICKET_TYPE_DEVILSQUARE);
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(1, byPossibleLevel);
+        return false;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 47)
+    {
+        const BYTE byPossibleLevel = CaculateFreeTicketLevel(FREETICKET_TYPE_BLOODCASTLE);
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(2, byPossibleLevel);
+        return false;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 48)
+    {
+        if (Hero->SafeZone || gMapManager.InHellas())
+        {
+            g_pSystemLogBox->AddText(GlobalText[1238], TYPE_ERROR_MESSAGE);
+            return false;
+        }
+
+        SendRequestUse(iIndex, 0);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 61)
+    {
+        const BYTE byPossibleLevel = CaculateFreeTicketLevel(FREETICKET_TYPE_CURSEDTEMPLE);
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(5, byPossibleLevel);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 121)
+    {
+        if (Hero->SafeZone == false)
+        {
+            CreateOkMessageBox(GlobalText[2330]);
+            return false;
+        }
+
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(4, pItem->Level);
+        g_pMyInventory->SetStandbyItemKey(pItem->Key);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_SCROLL_OF_BLOOD)
+    {
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(5, pItem->Level);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_DEVILS_INVITATION)
+    {
+        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(1, pItem->Level);
+        return true;
+    }
+
+    if (pItem->Type == ITEM_INVISIBILITY_CLOAK)
+    {
+        if (pItem->Level == 0)
+        {
+            g_pSystemLogBox->AddText(GlobalText[2089], TYPE_ERROR_MESSAGE);
+        }
+        else
+        {
+            SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(2, pItem->Level - 1);
+        }
+
+        return true;
+    }
+
+    if ((pItem->Type >= ITEM_SCROLL_OF_POISON && pItem->Type < ITEM_ETC + MAX_ITEM_INDEX)
+        || (pItem->Type >= ITEM_ORB_OF_TWISTING_SLASH && pItem->Type <= ITEM_ORB_OF_GREATER_FORTITUDE)
+        || (pItem->Type >= ITEM_ORB_OF_FIRE_SLASH && pItem->Type <= ITEM_ORB_OF_DEATH_STAB)
+        || (pItem->Type == ITEM_WING + 20)
+        || (pItem->Type >= ITEM_SCROLL_OF_FIREBURST && pItem->Type <= ITEM_SCROLL_OF_ELECTRIC_SPARK)
+        || (pItem->Type == ITEM_SCROLL_OF_FIRE_SCREAM)
+        || (pItem->Type == ITEM_CRYSTAL_OF_DESTRUCTION)
+        || (pItem->Type == ITEM_CRYSTAL_OF_FLAME_STRIKE)
+        || (pItem->Type == ITEM_CRYSTAL_OF_RECOVERY)
+        || (pItem->Type == ITEM_CRYSTAL_OF_MULTI_SHOT)
+        || (pItem->Type == ITEM_SCROLL_OF_CHAOTIC_DISEIER)
+        || (pItem->Type == ITEM_SCROLL_OF_GIGANTIC_STORM)
+        || (pItem->Type == ITEM_SCROLL_OF_WIZARDRY_ENHANCE))
+    {
+        bool bReadBookGem = true;
+
+        if (pItem->Type == ITEM_SCROLL_OF_NOVA
+            || pItem->Type == ITEM_SCROLL_OF_WIZARDRY_ENHANCE
+            || pItem->Type == ITEM_CRYSTAL_OF_MULTI_SHOT
+            || pItem->Type == ITEM_CRYSTAL_OF_RECOVERY
+            || pItem->Type == ITEM_CRYSTAL_OF_DESTRUCTION)
+        {
+            if (g_csQuest.getQuestState2(QUEST_CHANGE_UP_3) != QUEST_END)
+            {
+                bReadBookGem = false;
+            }
+        }
+
+        if (pItem->Type == ITEM_SCROLL_OF_CHAOTIC_DISEIER)
+        {
+            if (CharacterAttribute->Level < 220)
+            {
+                bReadBookGem = false;
+            }
+        }
+
+        if (bReadBookGem)
+        {
+            const WORD wStrength = CharacterAttribute->Strength + CharacterAttribute->AddStrength;
+            const WORD wEnergy   = CharacterAttribute->Energy   + CharacterAttribute->AddEnergy;
+
+            if (CharacterAttribute->Level >= ItemAttribute[pItem->Type].RequireLevel
+                && wEnergy   >= pItem->RequireEnergy
+                && wStrength >= pItem->RequireStrength)
+            {
+                SendRequestUse(iIndex, 0);
+            }
+
+            return true;
+        }
+
+        return false;
+    }
+
+    if (pItem->Type == ITEM_FRUITS)
+    {
+        if (CharacterAttribute->Level < 10)
+        {
+            CreateOkMessageBox(GlobalText[749]);
+            return true;
+        }
+
+        bool bEquipmentEmpty = true;
+        for (int i = 0; i < MAX_EQUIPMENT; i++)
+        {
+            if (CharacterMachine->Equipment[i].Type != -1)
+            {
+                bEquipmentEmpty = false;
+                break;
+            }
+        }
+
+        if (!bEquipmentEmpty)
+        {
+            CreateOkMessageBox(GlobalText[1909]);
+            return true;
+        }
+
+        if (pItem->Level == 4) // Command Fruit
+        {
+            if (gCharacterManager.GetBaseClass(CharacterAttribute->Class) != CLASS_DARK_LORD)
+            {
+                CreateOkMessageBox(GlobalText[1905]);
+                return true;
+            }
+        }
+
+        g_pMyInventory->SetStandbyItemKey(pItem->Key);
+        CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUseFruitMsgBoxLayout));
+        return true;
+    }
+
+    if (pItem->Type == ITEM_LIFE_STONE_ITEM)
+    {
+        bool bUse = false;
+        switch (pItem->Level)
+        {
+        case 0: bUse = true; break;
+        case 1: if (Hero->GuildStatus != G_MASTER) bUse = true; break;
+        }
+
+        if (bUse)
+        {
+            SendRequestUse(iIndex, 0);
+            return true;
+        }
+
+        return false;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 69)
+    {
+        if (g_PortalMgr.IsRevivePositionSaved())
+        {
+            if (g_PortalMgr.IsPortalUsable())
+            {
+                g_pMyInventory->SetStandbyItemKey(pItem->Key);
+                CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUseReviveCharmMsgBoxLayout));
+            }
+            else
+            {
+                CreateOkMessageBox(GlobalText[2608]);
+            }
+        }
+
+        return false;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 70)
+    {
+        if (g_PortalMgr.IsPortalUsable())
+        {
+            if (pItem->Durability == 2)
+            {
+                if (g_PortalMgr.IsPortalPositionSaved())
+                {
+                    CreateOkMessageBox(GlobalText[2610]);
+                }
+                else
+                {
+                    g_pMyInventory->SetStandbyItemKey(pItem->Key);
+                    CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUsePortalCharmMsgBoxLayout));
+                }
+            }
+            else if (pItem->Durability == 1)
+            {
+                g_pMyInventory->SetStandbyItemKey(pItem->Key);
+                CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CReturnPortalCharmMsgBoxLayout));
+            }
+        }
+        else
+        {
+            CreateOkMessageBox(GlobalText[2608]);
+        }
+
+        return false;
+    }
+
+    if (pItem->Type == ITEM_HELPER + 66)
+    {
+        g_pMyInventory->SetStandbyItemKey(pItem->Key);
+        CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUseSantaInvitationMsgBoxLayout));
+    }
+
+    return false;
+}
+
+bool CNewUIInventoryActionController::TryTransferBetweenInventorySections(CNewUIInventoryCtrl* sourceControl) const
+{
+    if (sourceControl == nullptr || g_pMyInventoryExt == nullptr)
+    {
+        return false;
+    }
+
+    ITEM* pItem = sourceControl->FindItemAtPt(MouseX, MouseY);
+    if (pItem == nullptr)
+    {
+        return false;
+    }
+
+    const int sourceIndex = sourceControl->GetIndexByItem(pItem);
+    if (sourceIndex < 0)
+    {
+        return false;
+    }
+
+    const ITEM_ATTRIBUTE* itemAttribute = &ItemAttribute[pItem->Type];
+    int destinationIndex = -1;
+
+    if (sourceControl == g_pMyInventory->GetInventoryCtrl())
+    {
+        destinationIndex = g_pMyInventoryExt->FindEmptySlot(
+            itemAttribute->Width, itemAttribute->Height);
+    }
+    else
+    {
+        destinationIndex = m_pContext->FindEmptySlot(pItem);
+    }
+
+    if (destinationIndex == -1 || destinationIndex == sourceIndex)
+    {
+        return true;
+    }
+
+    if (!CNewUIInventoryCtrl::CreatePickedItem(sourceControl, pItem))
+    {
+        return false;
+    }
+
+    sourceControl->RemoveItem(pItem);
+
+    CNewUIPickedItem* pPickedItem = CNewUIInventoryCtrl::GetPickedItem();
+    if (pPickedItem == nullptr || pPickedItem->GetItem() == nullptr)
+    {
+        CNewUIInventoryCtrl::BackupPickedItem();
+        return false;
+    }
+
+    pPickedItem->HidePickedItem();
+
+    if (!SendRequestEquipmentItem(STORAGE_TYPE::INVENTORY, sourceIndex,
+        pPickedItem->GetItem(), STORAGE_TYPE::INVENTORY, destinationIndex))
+    {
+        CNewUIInventoryCtrl::BackupPickedItem();
+        return false;
+    }
 
     return true;
 }
