+++ src/source/NewUIInventoryActionController.h
@@ -1,4 +1,4 @@
-// NewUIInventoryActionController.h: interface for the CNewUIInventoryActionController class.
+// NewUIInventoryActionController.h
 //////////////////////////////////////////////////////////////////////
 
 #if !defined(AFX_NEWUIINVENTORYACTIONCONTROLLER_H__INCLUDED_)
@@ -7,11 +7,10 @@
 #pragma once
 
 #include "NewUIInventoryCtrl.h"
+#include "IInventoryActionContext.h"
 
 namespace SEASON3B
 {
-
-class CNewUIMyInventory;
 
 class CNewUIInventoryActionController
 {
@@ -19,17 +18,14 @@ public:
     CNewUIInventoryActionController();
     ~CNewUIInventoryActionController() = default;
 
-    void SetOwner(CNewUIMyInventory* pOwner);
+    void SetContext(IInventoryActionContext* pContext);
     bool HandleInventoryActions(CNewUIInventoryCtrl* targetControl) const;
 
 private:
     bool HandlePickedItemPlacement(CNewUIInventoryCtrl* targetControl) const;
-    bool TryApplyJewel(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem,
-                       ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
-    bool TryStackItem(CNewUIInventoryCtrl* targetControl, ITEM* pPickItem,
-                      int iSourceIndex, int iTargetIndex) const;
-    bool TryMoveItem(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem,
-                     ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
+    bool TryApplyJewel(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
+    bool TryStackItem(CNewUIInventoryCtrl* targetControl, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
+    bool TryMoveItem(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
 
     bool HandleRepairClick(CNewUIInventoryCtrl* targetControl) const;
 
@@ -40,10 +36,16 @@ private:
     bool TryEquipItem(CNewUIInventoryCtrl* targetControl, ITEM* pItem, int iSrcIndex) const;
     bool TryDropItem(CNewUIInventoryCtrl* targetControl, ITEM* pItem) const;
 
-    int FindAlternateEquipSlot(int nOriginalSlot, ITEM* pItem) const;
+    int  FindAlternateEquipSlot(int nOriginalSlot, ITEM* pItem) const;
     bool IsSlotOccupied(int nSlot) const;
 
-    CNewUIMyInventory* m_pOwner;
+    bool ApplyJewels(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
+    bool TryStackItems(CNewUIInventoryCtrl* targetControl, ITEM* pPickItem, int iSourceIndex, int iTargetIndex) const;
+    bool RepairItemAtMousePoint(CNewUIInventoryCtrl* targetControl) const;
+    bool TryConsumeItem(CNewUIInventoryCtrl* targetControl, ITEM* pItem, int iIndex) const;
+    bool TryTransferBetweenInventorySections(CNewUIInventoryCtrl* sourceControl) const;
+
+    IInventoryActionContext* m_pContext;
 };
 
 } // namespace SEASON3B
