+++ src/source/NewUIMyInventory.cpp
@@ -73,7 +73,7 @@ bool CNewUIMyInventory::Create(CNewUIManager* pNewUIMng, CNewUI3DRenderMng* pNew
         return false;
     }
 
-    m_ActionController.SetOwner(this);
+    m_ActionController.SetContext(this); 
 
     SetPos(x, y);
     LoadImages();
@@ -199,7 +199,7 @@ void CNewUIMyInventory::UnequipAllItems()
     }
 }
 
-bool CNewUIMyInventory::IsEquipable(int iIndex, ITEM* pItem)
+bool CNewUIMyInventory::IsEquipable(int iIndex, ITEM* pItem) const
 {
     if (pItem == nullptr)
         return false;
@@ -427,7 +427,7 @@ const POINT& CNewUIMyInventory::GetPos() const
     return m_Pos;
 }
 
-CNewUIMyInventory::REPAIR_MODE CNewUIMyInventory::GetRepairMode() const
+SEASON3B::REPAIR_MODE CNewUIMyInventory::GetRepairMode() const
 {
     return m_RepairMode;
 }
@@ -1550,585 +1550,6 @@ bool CNewUIMyInventory::EquipmentWindowProcess()
     }
 
     return false;
-}
-
-bool CNewUIMyInventory::RepairItemAtMousePoint(CNewUIInventoryCtrl* targetControl)
-{
-    ITEM* pItem = targetControl->FindItemAtPt(MouseX, MouseY);
-    if (pItem == nullptr)
-    {
-        return true;
-    }
-
-    if (IsRepairBan(pItem) == true)
-    {
-        return true;
-    }
-
-    int iIndex = targetControl->GetIndex(pItem->x, pItem->y);
-
-    if (g_pNewUISystem->IsVisible(INTERFACE_NPCSHOP) && g_pNPCShop->IsRepairShop())
-    {
-        SocketClient->ToGameServer()->SendRepairItemRequest(iIndex, 0);
-    }
-    else
-    {
-        SocketClient->ToGameServer()->SendRepairItemRequest(iIndex, 1);
-    }
-
-    return true;
-}
-
-bool CNewUIMyInventory::ApplyJewels(CNewUIInventoryCtrl* targetControl, CNewUIPickedItem* pPickedItem, ITEM* pPickItem, const int iSourceIndex, const int iTargetIndex)
-{
-    auto pItem = targetControl->FindItem(iTargetIndex);
-    if (!pItem)
-    {
-        return false;
-    }
-
-    const int	iType = pItem->Type;
-    const int	iLevel = pItem->Level;
-    const int	iDurability = pItem->Durability;
-
-    bool bSuccess = true;
-
-    if (iType > ITEM_WINGS_OF_DARKNESS
-        && iType != ITEM_CAPE_OF_LORD
-        && !(iType >= ITEM_WING_OF_STORM && iType <= ITEM_WING_OF_DIMENSION)
-        && !(ITEM_WING + 130 <= iType && iType <= ITEM_WING + 134)
-        && !(iType >= ITEM_CAPE_OF_FIGHTER && iType <= ITEM_CAPE_OF_OVERRULE)
-        && (iType != ITEM_WING + 135))
-    {
-        bSuccess = false;
-    }
-
-    if (iType == ITEM_BOLT || iType == ITEM_ARROWS)
-    {
-        bSuccess = false;
-    }
-
-    if ((pPickItem->Type == ITEM_JEWEL_OF_BLESS && iLevel >= 6) || (pPickItem->Type == ITEM_JEWEL_OF_SOUL && iLevel >= 9))
-    {
-        bSuccess = false;
-    }
-
-    if (pPickItem->Type == ITEM_JEWEL_OF_BLESS && iType == ITEM_HORN_OF_FENRIR && iDurability != 255)
-    {
-        CFenrirRepairMsgBox* pMsgBox = nullptr;
-        CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CFenrirRepairMsgBoxLayout), &pMsgBox);
-        pMsgBox->SetSourceIndex(iSourceIndex);
-
-        auto iIndex = targetControl->GetIndex(pItem->x, pItem->y);
-        pMsgBox->SetTargetIndex(iIndex);
-
-        pPickedItem->HidePickedItem();
-        return true;
-    }
-
-    if (pPickItem->Type == ITEM_JEWEL_OF_HARMONY)
-    {
-        if (g_SocketItemMgr.IsSocketItem(pItem))
-        {
-            bSuccess = false;
-        }
-        else
-            if (pItem->Jewel_Of_Harmony_Option != 0)
-            {
-                bSuccess = false;
-            }
-            else
-            {
-                const StrengthenItem strengthitem = g_pUIJewelHarmonyinfo->GetItemType(static_cast<int>(pItem->Type));
-
-                if (strengthitem == SI_None)
-                {
-                    bSuccess = false;
-                }
-            }
-    }
-
-    if (pPickItem->Type == ITEM_LOWER_REFINE_STONE || pPickItem->Type == ITEM_HIGHER_REFINE_STONE)
-    {
-        if (g_SocketItemMgr.IsSocketItem(pItem))
-        {
-            bSuccess = false;
-        }
-        else if (pItem->Jewel_Of_Harmony_Option == 0)
-        {
-            bSuccess = false;
-        }
-    }
-
-    if (Check_LuckyItem(pItem->Type))
-    {
-        bSuccess = false;
-        if (pPickItem->Type == ITEM_POTION + 161)
-        {
-            if (pItem->Jewel_Of_Harmony_Option == 0)	bSuccess = true;
-        }
-        else if (pPickItem->Type == ITEM_POTION + 160)
-        {
-            if (pItem->Durability > 0)					bSuccess = true;
-        }
-    }
-
-    if (bSuccess)
-    {
-        const int targetIndex = targetControl->GetIndexByItem(pItem);
-        SendRequestUse(iSourceIndex, targetIndex);
-        PlayBuffer(SOUND_GET_ITEM01);
-        return true;
-    }
-
-    return false;
-}
-
-bool CNewUIMyInventory::TryStackItems(CNewUIInventoryCtrl* targetControl, ITEM* pPickItem, const int iSourceIndex, const int iTargetIndex)
-{
-    if (ITEM* pItem = targetControl->FindItem(iTargetIndex))
-    {
-        if (targetControl->AreItemsStackable(pPickItem, pItem))
-        {
-            SendRequestEquipmentItem(STORAGE_TYPE::INVENTORY, iSourceIndex, pPickItem, STORAGE_TYPE::INVENTORY, iTargetIndex);
-            return true;
-        }
-    }
-
-    return false;
-}
-
-bool CNewUIMyInventory::TryConsumeItem(CNewUIInventoryCtrl* targetControl, ITEM* pItem, const int iIndex)
-{
-    if (pItem == nullptr)
-    {
-        return false;
-    }
-
-    if (pItem->Type == ITEM_TOWN_PORTAL_SCROLL)
-    {
-        SendRequestUse(iIndex, 0);
-        return true;
-    }
-
-    const auto isApple = pItem->Type == ITEM_APPLE;
-    const auto isPotion =
-        (pItem->Type >= ITEM_APPLE && pItem->Type <= ITEM_ALE)
-        || (pItem->Type >= ITEM_SMALL_SHIELD_POTION && pItem->Type <= ITEM_LARGE_COMPLEX_POTION);
-
-    if (isApple || isPotion
-        || (pItem->Type == ITEM_POTION + 20 && pItem->Level == 0)
-        || (pItem->Type >= ITEM_JACK_OLANTERN_BLESSINGS && pItem->Type <= ITEM_JACK_OLANTERN_DRINK)
-        || (pItem->Type == ITEM_BOX_OF_LUCK && pItem->Level == 14)
-        || (pItem->Type >= ITEM_POTION + 70 && pItem->Type <= ITEM_POTION + 71)
-        || (pItem->Type >= ITEM_POTION + 72 && pItem->Type <= ITEM_POTION + 77)
-        || pItem->Type == ITEM_HELPER + 60
-        || pItem->Type == ITEM_POTION + 94
-        || (pItem->Type >= ITEM_CHERRY_BLOSSOM_WINE && pItem->Type <= ITEM_CHERRY_BLOSSOM_FLOWER_PETAL)
-        || (pItem->Type >= ITEM_POTION + 97 && pItem->Type <= ITEM_POTION + 98)
-        || pItem->Type == ITEM_HELPER + 81
-        || pItem->Type == ITEM_HELPER + 82
-        || pItem->Type == ITEM_POTION + 133
-        )
-    {
-        SendRequestUse(iIndex, 0);
-        if (isApple)
-        {
-            PlayBuffer(SOUND_EAT_APPLE01);
-        }
-        else if (isPotion)
-        {
-            PlayBuffer(SOUND_DRINK01);
-        }
-
-        return true;
-    }
-
-    if ((pItem->Type >= ITEM_POTION + 78 && pItem->Type <= ITEM_POTION + 82))
-    {
-        std::list<eBuffState> secretPotionbufflist;
-        secretPotionbufflist.push_back(eBuff_SecretPotion1);
-        secretPotionbufflist.push_back(eBuff_SecretPotion2);
-        secretPotionbufflist.push_back(eBuff_SecretPotion3);
-        secretPotionbufflist.push_back(eBuff_SecretPotion4);
-        secretPotionbufflist.push_back(eBuff_SecretPotion5);
-
-        if (g_isCharacterBufflist((&Hero->Object), secretPotionbufflist) == eBuffNone) {
-            SendRequestUse(iIndex, 0);
-            return true;
-        }
-
-        CreateOkMessageBox(GlobalText[2530], RGBA(255, 30, 0, 255));
-        return false;
-    }
-
-    if ((pItem->Type >= ITEM_HELPER + 54 && pItem->Type <= ITEM_HELPER + 57) || (pItem->Type == ITEM_HELPER + 58 && gCharacterManager.GetBaseClass(Hero->Class) == CLASS_DARK_LORD))
-    {
-        bool result = true;
-        WORD point[5] = { 0, };
-
-        point[0] = CharacterAttribute->Strength + CharacterAttribute->AddStrength;
-        point[1] = CharacterAttribute->Dexterity + CharacterAttribute->AddDexterity;
-        point[2] = CharacterAttribute->Vitality + CharacterAttribute->AddVitality;
-        point[3] = CharacterAttribute->Energy + CharacterAttribute->AddEnergy;
-        point[4] = CharacterAttribute->Charisma + CharacterAttribute->AddCharisma;
-
-        wchar_t nStat[MAX_CLASS][5] =
-        {
-            18, 18, 15, 30,	0,
-            28, 20, 25, 10,	0,
-            22, 25, 20, 15,	0,
-            26, 26, 26, 26,	0,
-            26, 20, 20, 15, 25,
-            21, 21, 18, 23,	0,
-            32, 27, 25, 20, 0,
-        };
-
-        const auto attributeType = pItem->Type - (ITEM_HELPER + 54);
-        const int characterClass = gCharacterManager.GetBaseClass(Hero->Class);
-        point[attributeType] -= nStat[characterClass][attributeType];
-        if (point[attributeType] < (pItem->Durability * 10))
-        {
-            g_pMyInventory->SetStandbyItemKey(pItem->Key);
-            CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUsePartChargeFruitMsgBoxLayout));
-            return false;
-        }
-
-        SendRequestUse(iIndex, 0);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 58 && gCharacterManager.GetBaseClass(Hero->Class) != CLASS_DARK_LORD)
-    {
-        CreateOkMessageBox(GlobalText[1905]);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_ARMOR_OF_GUARDSMAN)
-    {
-        if (IsUnitedMarketPlace())
-        {
-            wchar_t	szOutputText[512];
-            mu_swprintf(szOutputText, L"%ls %ls", GlobalText[3014], GlobalText[3015]);
-
-            CreateOkMessageBox(szOutputText);
-            return true;
-        }
-
-        if (Hero->SafeZone == false)
-        {
-            CreateOkMessageBox(GlobalText[2330]);
-            return false;
-        }
-
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(4, pItem->Level);
-        g_pMyInventory->SetStandbyItemKey(pItem->Key);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 46)
-    {
-        const BYTE byPossibleLevel = CaculateFreeTicketLevel(FREETICKET_TYPE_DEVILSQUARE);
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(1, byPossibleLevel);
-        return false;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 47)
-    {
-        const BYTE byPossibleLevel = CaculateFreeTicketLevel(FREETICKET_TYPE_BLOODCASTLE);
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(2, byPossibleLevel);
-        return false;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 48)
-    {
-        if (Hero->SafeZone || gMapManager.InHellas())
-        {
-            g_pSystemLogBox->AddText(GlobalText[1238], TYPE_ERROR_MESSAGE);
-            return false;
-        }
-
-        SendRequestUse(iIndex, 0);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 61)
-    {
-        const BYTE byPossibleLevel = CaculateFreeTicketLevel(FREETICKET_TYPE_CURSEDTEMPLE);
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(5, byPossibleLevel);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 121)
-    {
-        if (Hero->SafeZone == false)
-        {
-            CreateOkMessageBox(GlobalText[2330]);
-            return false;
-        }
-
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(4, pItem->Level);
-        g_pMyInventory->SetStandbyItemKey(pItem->Key);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_SCROLL_OF_BLOOD)
-    {
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(5, pItem->Level);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_DEVILS_INVITATION)
-    {
-        SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(1, pItem->Level);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_INVISIBILITY_CLOAK)
-    {
-        if (pItem->Level == 0)
-        {
-            g_pSystemLogBox->AddText(GlobalText[2089], TYPE_ERROR_MESSAGE);
-        }
-        else
-        {
-            SocketClient->ToGameServer()->SendMiniGameOpeningStateRequest(2, pItem->Level - 1);
-        }
-
-        return true;
-    }
-
-    if ((pItem->Type >= ITEM_SCROLL_OF_POISON && pItem->Type < ITEM_ETC + MAX_ITEM_INDEX)
-        || (pItem->Type >= ITEM_ORB_OF_TWISTING_SLASH && pItem->Type <= ITEM_ORB_OF_GREATER_FORTITUDE)
-        || (pItem->Type >= ITEM_ORB_OF_FIRE_SLASH && pItem->Type <= ITEM_ORB_OF_DEATH_STAB)
-        || (pItem->Type == ITEM_WING + 20)
-        || (pItem->Type >= ITEM_SCROLL_OF_FIREBURST && pItem->Type <= ITEM_SCROLL_OF_ELECTRIC_SPARK)
-        || (pItem->Type == ITEM_SCROLL_OF_FIRE_SCREAM)
-        || (pItem->Type == ITEM_CRYSTAL_OF_DESTRUCTION)
-        || (pItem->Type == ITEM_CRYSTAL_OF_FLAME_STRIKE)
-        || (pItem->Type == ITEM_CRYSTAL_OF_RECOVERY)
-        || (pItem->Type == ITEM_CRYSTAL_OF_MULTI_SHOT)
-        || (pItem->Type == ITEM_SCROLL_OF_CHAOTIC_DISEIER)
-        || (pItem->Type == ITEM_SCROLL_OF_GIGANTIC_STORM)
-        || (pItem->Type == ITEM_SCROLL_OF_WIZARDRY_ENHANCE)
-        )
-    {
-        bool bReadBookGem = true;
-
-        if ((pItem->Type == ITEM_SCROLL_OF_NOVA)
-            || (pItem->Type == ITEM_SCROLL_OF_WIZARDRY_ENHANCE)
-            || pItem->Type == ITEM_CRYSTAL_OF_MULTI_SHOT
-            || (pItem->Type == ITEM_CRYSTAL_OF_RECOVERY)
-            || (pItem->Type == ITEM_CRYSTAL_OF_DESTRUCTION)
-            )
-        {
-            if (g_csQuest.getQuestState2(QUEST_CHANGE_UP_3) != QUEST_END)
-
-                bReadBookGem = false;
-        }
-        if (pItem->Type == ITEM_SCROLL_OF_CHAOTIC_DISEIER)
-        {
-            const int Level = CharacterAttribute->Level;
-            if (Level < 220)
-                bReadBookGem = false;
-        }
-        if (bReadBookGem)
-        {
-            WORD Strength, Energy;
-            Strength = CharacterAttribute->Strength + CharacterAttribute->AddStrength;
-            Energy = CharacterAttribute->Energy + CharacterAttribute->AddEnergy;
-
-            if (CharacterAttribute->Level >= ItemAttribute[pItem->Type].RequireLevel &&
-                Energy >= pItem->RequireEnergy &&
-                Strength >= pItem->RequireStrength)
-            {
-                SendRequestUse(iIndex, 0);
-            }
-
-            return true;
-        }
-
-        return false;
-    }
-
-    if (pItem->Type == ITEM_FRUITS)
-    {
-        const int Level = CharacterAttribute->Level;
-
-        if (Level >= 10)
-        {
-            bool bEquipmentEmpty = true;
-            for (int i = 0; i < MAX_EQUIPMENT; i++)
-            {
-                if (CharacterMachine->Equipment[i].Type != -1)
-                {
-                    bEquipmentEmpty = false;
-                }
-            }
-            const auto Class = CharacterAttribute->Class;
-
-            if (bEquipmentEmpty == true)
-            {
-                if (pItem->Level == 4) // Command Fruit
-                {
-                    if (gCharacterManager.GetBaseClass(Class) != CLASS_DARK_LORD)
-                    {
-                        CreateOkMessageBox(GlobalText[1905]);
-                        return true;
-                    }
-                }
-                g_pMyInventory->SetStandbyItemKey(pItem->Key);
-                CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUseFruitMsgBoxLayout));
-                return true;
-            }
-
-            CreateOkMessageBox(GlobalText[1909]);
-            return true;
-        }
-
-        CreateOkMessageBox(GlobalText[749]);
-        return true;
-    }
-
-    if (pItem->Type == ITEM_LIFE_STONE_ITEM)
-    {
-        bool bUse = false;
-        switch (pItem->Level)
-        {
-        case 0:
-            bUse = true;
-            break;
-        case 1:
-            if (Hero->GuildStatus != G_MASTER)
-                bUse = true;
-            break;
-        }
-        if (bUse)
-        {
-            SendRequestUse(iIndex, 0);
-            return true;
-        }
-
-        return false;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 69)
-    {
-        if (g_PortalMgr.IsRevivePositionSaved())
-        {
-            if (g_PortalMgr.IsPortalUsable())
-            {
-                g_pMyInventory->SetStandbyItemKey(pItem->Key);
-                CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUseReviveCharmMsgBoxLayout));
-            }
-            else
-            {
-                CreateOkMessageBox(GlobalText[2608]);
-            }
-        }
-
-        return false;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 70)
-    {
-        if (g_PortalMgr.IsPortalUsable())
-        {
-            if (pItem->Durability == 2)
-            {
-                if (g_PortalMgr.IsPortalPositionSaved())
-                {
-                    CreateOkMessageBox(GlobalText[2610]);
-                }
-                else
-                {
-                    g_pMyInventory->SetStandbyItemKey(pItem->Key);
-                    CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUsePortalCharmMsgBoxLayout));
-                }
-            }
-            else if (pItem->Durability == 1)
-            {
-                g_pMyInventory->SetStandbyItemKey(pItem->Key);
-                CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CReturnPortalCharmMsgBoxLayout));
-            }
-        }
-        else
-        {
-            CreateOkMessageBox(GlobalText[2608]);
-        }
-
-        return false;
-    }
-
-    if (pItem->Type == ITEM_HELPER + 66)
-    {
-        g_pMyInventory->SetStandbyItemKey(pItem->Key);
-        CreateMessageBox(MSGBOX_LAYOUT_CLASS(SEASON3B::CUseSantaInvitationMsgBoxLayout));
-    }
-
-    return false;
-}
-
-bool CNewUIMyInventory::TryTransferBetweenInventorySections(CNewUIInventoryCtrl* sourceControl)
-{
-    if (sourceControl == nullptr || g_pMyInventoryExt == nullptr)
-    {
-        return false;
-    }
-
-    ITEM* pItem = sourceControl->FindItemAtPt(MouseX, MouseY);
-    if (pItem == nullptr)
-    {
-        return false;
-    }
-
-    const int sourceIndex = sourceControl->GetIndexByItem(pItem);
-    if (sourceIndex < 0)
-    {
-        return false;
-    }
-
-    const ITEM_ATTRIBUTE* itemAttribute = &ItemAttribute[pItem->Type];
-    int destinationIndex = -1;
-
-    if (sourceControl == g_pMyInventory->GetInventoryCtrl())
-    {
-        destinationIndex = g_pMyInventoryExt->FindEmptySlot(itemAttribute->Width, itemAttribute->Height);
-    }
-    else
-    {
-        destinationIndex = g_pMyInventory->FindEmptySlot(itemAttribute->Width, itemAttribute->Height);
-    }
-
-    if (destinationIndex == -1 || destinationIndex == sourceIndex)
-    {
-        return true;
-    }
-
-    if (!CNewUIInventoryCtrl::CreatePickedItem(sourceControl, pItem))
-    {
-        return false;
-    }
-
-    sourceControl->RemoveItem(pItem);
-
-    CNewUIPickedItem* pickedItem = CNewUIInventoryCtrl::GetPickedItem();
-    if (pickedItem == nullptr || pickedItem->GetItem() == nullptr)
-    {
-        CNewUIInventoryCtrl::BackupPickedItem();
-        return false;
-    }
-
-    pickedItem->HidePickedItem();
-
-    if (!SendRequestEquipmentItem(STORAGE_TYPE::INVENTORY, sourceIndex,
-        pickedItem->GetItem(), STORAGE_TYPE::INVENTORY, destinationIndex))
-    {
-        CNewUIInventoryCtrl::BackupPickedItem();
-        return false;
-    }
-
-    return true;
 }
 
 bool CNewUIMyInventory::InventoryProcess() const
